Remove validation set from -FrameworkVersion parameter
The user gets to specify the version of NUNit they want to use. Different versions of NUnit support different frameworks, and are more than happy to complain if an unsupported framework version is specified. There's no need to apply our own validation, and it's actually got in the way a number of times recently.